+++ src/test/java/org/mule/module/hubspot/integration/utils/CloneGenerator.java
@@ -1,0 +1,44 @@
+/**
+ *
+ * (c) 2003-2012 MuleSoft, Inc. This software is protected under international
+ * copyright law. All use of this software is subject to MuleSoft's Master
+ * Subscription Agreement (or other Terms of Service) separately entered
+ * into between you and MuleSoft. If such an agreement is not in
+ * place, you may not use the software.
+ */
+
+package org.mule.module.hubspot.integration.utils;
+
+import java.lang.reflect.Field;
+
+public class CloneGenerator {
+
+	public static Object clone(Object o) {
+		Object clone = null;
+
+		try {
+			clone = o.getClass().newInstance();
+		} catch (InstantiationException e) {
+			e.printStackTrace();
+		} catch (IllegalAccessException e) {
+			e.printStackTrace();
+		}
+
+		// Walk up the superclass hierarchy
+		for (Class<?> obj = o.getClass(); !obj.equals(Object.class); obj = obj.getSuperclass()) {
+			Field[] fields = obj.getDeclaredFields();
+			for (int i = 0; i < fields.length; i++) {
+				fields[i].setAccessible(true);
+				try {
+					// for each class/suerclass, copy all fields
+					// from this object to the clone
+					fields[i].set(clone, fields[i].get(o));
+				} catch (IllegalArgumentException e) {
+				} catch (IllegalAccessException e) {
+				}
+			}
+		}
+
+		return clone;
+	}
+}